Changes in "Soldier.run()"

+++ src/Game/Model/Buildings/Building.java
@@ -8,7 +8,7 @@ public abstract class Building extends Fightable implements Card {
     protected final int cost;
 
 
-    public Building(Board board, int hp, int damage, double hitSpeed, double range, Target target, int lifeTime,
+    public Building(Board board, int hp, int damage, long hitSpeed, double range, Target target, int lifeTime,
                     int cost, Location location , Team team) {
         super(board, hp, damage, hitSpeed, range, location, team , Type.BUILDING);
         this.target = target;
